fix(consumer): version pending clear markers
Allow only the exact marker captured before topic recreation so later ABA revocations or divergence corrections cannot be mistaken for stale state.

+++ tests/Dekaf.Tests.Unit/Consumer/ConsumerDiagnosticSnapshotTests.cs
@@ -31,7 +31,7 @@ public sealed class ConsumerDiagnosticSnapshotTests
         consumer.IncrementalAssign(
             [new TopicPartitionOffset(partition.Topic, partition.Partition, 42)]);
 
-        GetField<ConcurrentDictionary<TopicPartition, byte>>(
+        GetField<ConcurrentDictionary<TopicPartition, long>>(
             consumer,
             "_coordinatorRevokedPartitionsPendingFetchClear")[partition] = 0;
         GetField<ConcurrentDictionary<TopicPartition, int>>(
